pyproject.toml
[project]
name = "catalog-api-studio"
version = "0.1.0"
description = "Local desktop app for digitizing supplier catalogs into structured data, REST API, and smart search"
requires-python = ">=3.11"
dependencies = [
    # UI
    "PySide6>=6.6",
    # API
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    # DB
    "SQLAlchemy>=2.0",
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
    # PDF
    "PyMuPDF>=1.24",
    "pdfplumber>=0.10",
    "Pillow>=10.0",
    # Import
    "openpyxl>=3.1",
    "pandas>=2.1",
    # OCR
    "paddleocr>=2.7",
    "paddlepaddle>=2.6",
    # Search
    "typesense>=0.20",
    # Existing deps
    "watchdog>=4.0",
    "flask>=3.0",
    "opencv-python-headless>=4.8",
    "numpy>=1.24",
    "beautifulsoup4>=4.12",
    "requests>=2.31",
    "lxml>=5.0",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "pytest",
    "pytest-asyncio",
    "httpx",
]

[project.scripts]
catalog-api-studio = "catalog_api_studio.main:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
